requirements.txt
elementpath==1.1.8
PyYAML==5.1.1
xmlschema==1.0.13
